+++ hnote/Screens/ListScreen.tsx
@@ -2,8 +2,13 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { View, Text, TextInput, Button, ScrollView } from 'react-native';
 import { addTaskToList, updateTask } from '../actions/index';
+import styled from 'styled-components';
 
 import SingleTask from '../components/molecules/SingleTask';
+
+const Wrapper = styled.View`
+    display: ${props => (props.isDetailOpen ? 'none' : 'flex')};
+`;
 
 const ListScreen = props => {
     const [enteredText, setEnteredText] = useState('');
@@ -15,26 +20,37 @@ const ListScreen = props => {
     const listID = props.navigation.getParam('listId');
     const allMyTasks = props.tasks.filter(task => task.list === listID);
     console.log(allMyTasks[0]);
+
+    const detailOpenHandler = () => {
+        props.dispatch(addTaskToList(listID, enteredText));
+    };
+
+    const sayHello = item => {
+        props.navigation.navigate({
+            routeName: 'TaskPage',
+            params: {
+                taskId: item.id,
+            },
+        });
+    };
     return (
         <View>
-            <TextInput onChangeText={inputChangeHandler}></TextInput>
-            <Button
-                title="Add"
-                onPress={() =>
-                    props.dispatch(addTaskToList(listID, enteredText))
-                }
-            />
-            <ScrollView>
-                {allMyTasks.map(item => (
-                    <SingleTask
-                        key={item.id}
-                        value={item.content}
-                        onChange={str =>
-                            props.dispatch(updateTask(item.id, str))
-                        }
-                    />
-                ))}
-            </ScrollView>
+            <Wrapper>
+                <TextInput onChangeText={inputChangeHandler}></TextInput>
+                <Button title="Add" onPress={() => detailOpenHandler()} />
+                <ScrollView>
+                    {allMyTasks.map(item => (
+                        <SingleTask
+                            key={item.id}
+                            value={item.content}
+                            // onChange={str =>
+                            //     props.dispatch(updateTask(item.id, str))
+                            // }
+                            onPress={item => sayHello(item)}
+                        />
+                    ))}
+                </ScrollView>
+            </Wrapper>
         </View>
     );
 };